Split cli package
cli now has cli.go for type CLI and command.go for type Command.
